Added missing expander to tables
